+++ app/controllers/services/summaries_controller.rb
@@ -17,7 +17,7 @@ class Services::SummariesController < Services::ApplicationController
     if @project_item.persisted?
       session.delete(session_key)
       session.delete(:selected_project)
-      redirect_to @project_item,
+      redirect_to project_service_path(@project_item.project, @project_item),
                   notice: "Service ordered sucessfully"
     else
       redirect_to service_configuration_path(@service),